backend/requirements.txt
fastapi==0.95.2 
pydantic==1.10.7
python-jose
motor
python-dotenv
firebase_admin
twilio
passlib
python-multipart
uvicorn
bcrypt
websockets